requirements.txt
# Dependencias principales del proyecto

# Google Cloud - Vertex AI (Gemini)
google-cloud-aiplatform==1.71.1

# Google Cloud - BigQuery
google-cloud-bigquery==3.26.0
db-dtypes==1.3.0

# Google Cloud - Core
google-auth==2.35.0
google-cloud-core==2.4.1

# Framework web (para el backend)
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.10.2
pydantic-settings==2.6.1

# Utilidades
python-dotenv==1.0.1
python-multipart==0.0.17

